minor changes on utils

+++ src/core/Utils.h
@@ -165,7 +165,7 @@ namespace kinski
     inline T map_value(const T &val, const T &src_min, const T &src_max,
                        const T &dst_min, const T &dst_max)
     {
-        float mix_val = clamp<float>((val - src_min) / (src_max - src_min), 0.f, 1.f);
+        float mix_val = clamp<float>((val - src_min) / (float)(src_max - src_min), 0.f, 1.f);
         return mix<T>(dst_min, dst_max, mix_val);
     }
 }
